--- Assets/Scripts/Highscores/NewHighscoreUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 436cdfce1d284544e9b17113cda1bde5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
