+++ Assets/UI Toolkit/HighscoreVisualTree.uxml
@@ -1,6 +1,7 @@
 <ui:UXML xmlns:ui="UnityEngine.UIElements" xmlns:uie="UnityEditor.UIElements" xsi="http://www.w3.org/2001/XMLSchema-instance" engine="UnityEngine.UIElements" editor="UnityEditor.UIElements" noNamespaceSchemaLocation="../../UIElementsSchema/UIElements.xsd" editor-extension-mode="False">
     <Style src="project://database/Assets/UI%20Toolkit/HighscoreStyle.uss?fileID=7433441132597879392&amp;guid=ea3e33529cbeebc40bf06b5e8e1fedf8&amp;type=3#HighscoreStyle" />
     <ui:VisualElement name="Background" style="flex-grow: 1; background-color: rgb(0, 0, 0); visibility: visible; display: flex; overflow: visible; align-items: center; justify-content: flex-end;">
+        <ui:ListView name="Highscores" fixed-item-height="20" show-border="false" style="height: 100%; width: 100%; background-color: rgb(255, 255, 255);" />
         <ui:Button text="Main Menu" display-tooltip-when-elided="true" name="MainMenu" class="text buttonText button" style="overflow: hidden;" />
     </ui:VisualElement>
 </ui:UXML>
