+++ Assets/UI Toolkit/MainMenuVisualTree.uxml
@@ -6,7 +6,7 @@
         </ui:VisualElement>
         <ui:VisualElement name="ButtonContainer" style="flex-direction: column; justify-content: space-between; align-items: center; min-width: auto; height: 585px;">
             <ui:Button text="Play" display-tooltip-when-elided="true" name="Play" class="text button buttonText" style="-unity-text-align: middle-center; white-space: normal;" />
-            <ui:Button text="Highscores" display-tooltip-when-elided="true" name="HighScores" class="text button buttonText" style="background-color: rgb(80, 231, 167);" />
+            <ui:Button text="Highscores" display-tooltip-when-elided="true" name="Highscores" class="text button buttonText" style="background-color: rgb(80, 231, 167);" />
             <ui:Button text="Tutorial" display-tooltip-when-elided="true" name="Tutorial" class="text button buttonText" style="background-color: rgb(172, 135, 201);" />
             <ui:Button text="Quit" display-tooltip-when-elided="true" name="Quit" class="text button buttonText" style="background-color: rgb(208, 50, 42);" />
         </ui:VisualElement>
